requirements-phase1.txt
# Core dependencies for Phase 1 services

# Async and networking
websockets==12.0
aiohttp==3.9.1

# Redis with async support (no aioredis needed!)
redis==5.0.1
hiredis==2.3.2

# Groq SDK
groq==0.4.1

# Google Drive integration
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Audio processing
pydub==0.25.1

# Environment and utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0